Individual post pages
Implemented content-single.php to display single posts in their entirety
on one page. Changed content.php to update The Loop on the "home" blog
page so that titles of posts lead to individual post pages. Also changed
"home" blog page so that each post preview shows just an excerpt, not
the full post.